refactor: Convert viewport context file to directory
