+++ packages/insight-viewer/src/Context/Viewport/const.ts
@@ -1,0 +1,12 @@
+import { Viewport } from './types'
+
+export const ViewportContextDefaultValue: Viewport = {
+  scale: 1,
+  invert: false,
+  hflip: false,
+  vflip: false,
+  x: 0,
+  y: 0,
+  windowWidth: 127,
+  windowCenter: 256,
+}
