+++ packages/insight-viewer/src/Context/Viewport/types.ts
@@ -1,0 +1,10 @@
+export interface Viewport {
+  scale: number
+  invert: boolean
+  hflip: boolean
+  vflip: boolean
+  x: number
+  y: number
+  windowWidth: number
+  windowCenter: number
+}
