+++ packages/insight-viewer/src/index.ts
@@ -1,2 +1,2 @@
 export { default } from './hooks/useInsightViewer'
-export type { Viewport } from './Context/Viewport'
+export type { Viewport } from './Context/Viewport/types'
